+++ ddlpy/ddlpy.py
@@ -35,7 +35,7 @@ logger = logging.getLogger(__name__)
 
 
 def _send_post_request(url, request, timeout=None):
-    logger.debug("requesting: {} with {}".format(url, json.dumps(request)))
+    logger.debug("requesting at {} with request: {}".format(url, json.dumps(request)))
     resp = requests.post(url, json=request, timeout=timeout)
     if not resp.ok:
         raise IOError("Request failed: {}".format(resp.text))