+++ HISTORY.rst
@@ -6,7 +6,7 @@ UNRELEASED
 ----------
 * avoid duplicated periods in dataframe returned by `ddlpy.measurements_amount()` in https://github.com/Deltares/ddlpy/pull/93
 * allow for different retrieval frequencies (including None) in `ddlpy.measurements()` in https://github.com/Deltares/ddlpy/pull/95
-* prevent silencing of error about the request being too large and code cleanups in https://github.com/Deltares/ddlpy/pull/97
+* only catch "Geen waarnemingen aanwezig!" error message and raise all others (for instance for a too large request) in https://github.com/Deltares/ddlpy/pull/97
 
 0.4.0 (2024-04-08)
 ------------------